+++ cvat-ui/webpack.config.js
@@ -1,4 +1,4 @@
-// Copyright (C) 2020-2021 Intel Corporation
+// Copyright (C) 2020-2022 Intel Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -146,7 +146,7 @@ module.exports = (env) => ({
         new CopyPlugin({
             patterns: [
                 {
-                    from: '../cvat-data/src/js/3rdparty/avc.wasm',
+                    from: '../cvat-data/src/ts/3rdparty/avc.wasm',
                     to: 'assets/3rdparty/',
                 },
             ],