+++ cvat/apps/profiler.py
@@ -1,7 +1,7 @@
 from django.apps import apps
 
 if apps.is_installed('silk'):
-    from silk.profiling.profiler import silk_profile
+    from silk.profiling.profiler import silk_profile  # pylint: disable=unused-import
 else:
     from functools import wraps
     def silk_profile(name=None):