+++ tests/rest_api/utils/__init__.py
@@ -1,3 +1,3 @@
-# Copyright (C) 2021 Intel Corporation
+# Copyright (C) 2021-2022 Intel Corporation
 #
 # SPDX-License-Identifier: MIT